requirements.txt
sentence-transformers
torch
scikit-learn
transformers
pdfplumber
python-docx

